creating object utill class
