+++ Util.js
@@ -1,4 +1,4 @@
-import {convertObjectToNestedStructure} from "./modules/convertToNestedObject";
+import {convertObjectToNestedStructure} from "./modules/objectUtil";
 export default class Util {
   /**
    * @param {*} value
@@ -26,8 +26,5 @@ export default class Util {
       (rv[x[key]] = rv[x[key]] || []).push(x);
       return rv;
     }, {});
-  }
-  static convertObjectToNestedStructure(object){
-    return convertObjectToNestedStructure(object)
   }
 }
